update regular expression search and search result

+++ SearchResult.java
@@ -1,6 +1,6 @@
 class SearchResult {
     public String result_string;
     public String[] result_array;
-    public double time_used;
+    public long time_used;
     public int count;
 }
